Return extension metadata on setup()
This allows the extension to declare it can handle parallel source file
reads safely.

For reference, we have been using this extension for many years in the
Python Spanish translation project with parallel source reading with no
issues. Since the extension doesn't declare this support, we have been
performing a manual registration (basically a copy of this extension's
setup(), see [1]).

[1] https://github.com/python/python-docs-es/blob/0dc7bbd9d9bd7e44e6ad39b4e2dffffeabbb3ab6/conf.py#L130-L142

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>